fix DTO id (int to long) because of Entity change

+++ src/main/java/org/cesi/fablab/api/repository/TypeOperationRepository.java
@@ -7,9 +7,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.transaction.annotation.Transactional;
 
 @Transactional(transactionManager = "transactionManager")
-public interface TypeOperationRepository extends JpaRepository<TypeOperationEntity, Integer> {
+public interface TypeOperationRepository extends JpaRepository<TypeOperationEntity, Long> {
     // @Query("SELECT t FROM Thing t WHERE t.fooIn = ?1")
     List<TypeOperationEntity> findByName(String name);
 
-    TypeOperationEntity findById(int id);
+    TypeOperationEntity findById(long id);
 }
